Adding matchState() function
